Attempt to fix build on amd64, see bug #1812008, reported by w_m_b - wmb1957. Do as suggested by Thomas R. - dtrauma.
git-svn-id: https://widelands.svn.sourceforge.net/svnroot/widelands/trunk@2828 37b2a8de-5219-0410-9f54-a31bc463ab9c

+++ src/fieldaction.cc
@@ -122,7 +122,8 @@ belongs to and trigger signal buildclicked.
 */
 void BuildGrid::clickslot(int32_t idx)
 {
-	buildclicked.call(reinterpret_cast<int32_t>(get_data(idx)));
+	buildclicked.call
+		(static_cast<int32_t>(reinterpret_cast<intptr_t>(get_data(idx))));
 }
 
 
@@ -136,7 +137,8 @@ building it belongs to and trigger signal buildmouseout.
 */
 void BuildGrid::mouseoutslot(int32_t idx)
 {
-	buildmouseout.call(reinterpret_cast<int32_t>(get_data(idx)));
+	buildmouseout.call
+		(static_cast<int32_t>(reinterpret_cast<intptr_t>(get_data(idx))));
 }
 
 
@@ -150,7 +152,8 @@ building it belongs to and trigger signal buildmousein.
 */
 void BuildGrid::mouseinslot(int32_t idx)
 {
-	buildmousein.call(reinterpret_cast<int32_t>(get_data(idx)));
+	buildmousein.call
+		(static_cast<int32_t>(reinterpret_cast<intptr_t>(get_data(idx))));
 }
